feat: lum-18 merge extractSubdomain with saas
